API DONE

+++ src/main/java/moe/src/leyline/interfaces/rest/CrudAPI.java
@@ -1,6 +1,7 @@
 package moe.src.leyline.interfaces.rest;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -9,7 +10,7 @@ import java.util.List;
 public interface CrudAPI<T extends Serializable> {
     public void update(String t) throws Exception;
     public void insert(String t) throws Exception;
-    public void delete(Integer id) throws Exception;
+    public void delete(BigInteger id) throws Exception;
     public List<T> list() throws Exception;
-    public T find(Integer id) throws Exception;
+    public T find(BigInteger id) throws Exception;
 }
